fix: workspace-wide clippy cleanup (Phase 2 IPC + workspace sweep)
- Fix IPC clone_on_copy and field_reassign_with_default warnings
- Fix 175 admin-ui clippy warnings (unused imports, dead code, closures, vars)
- Fix 214 synvoid-mesh warnings (field_reassign, borrow, too_many_arguments)
- Fix 37 synvoid-http warnings (type_complexity, too_many_arguments, borrow)
- Fix 8 synvoid-block-store warnings (collapsible_if, len_zero, field_reassign)
- Fix 10 synvoid-proxy warnings (manual_contains, field_reassign)
- Fix 5 synvoid-http-client warnings (bool_assert_comparison)
- Fix 4 synvoid-geoip warnings (field_reassign)
- Fix 1 synvoid-upstream warning (nonminimal_bool)
- Fix 1 synvoid-serverless warning (unused mut)
- Fix 1 synvoid-http3 warning (unused variable)
- Fix 25+ root synvoid crate warnings (doc, sort, return, borrow, etc.)
- Add reason comments to all new #[allow] attributes per WS4 lint policy
- Document validation results in plans/milestone_d_phase_2_validation.md

Workspace clippy: 0 errors (default features)
Workspace fmt: clean
Workspace check: 0 errors
147 files changed, 902 insertions, 882 deletions

+++ crates/synvoid-http/src/waf_decision.rs
@@ -170,7 +170,7 @@ where
             let cookie = format_secure_http_only_cookie(
                 &session_cookie_name,
                 &session_cookie_value,
-                session_cookie_max_age as u64,
+                session_cookie_max_age,
             );
             on_challenged(html.len() as u64);
             on_log(200, elapsed_ms());
